fix(kinesis): remove Kinesis aggregation size and count limits
Substation uses the `internal/aggregate` package to buffer events
with consideration to buffer, size, counts, and durations which
shadow the checks here with adherence to Kinesis limits in a more
user friendly way. Inclusion of the size checks here allowed events
between 25KiB and 1MiB to be sent as empty events to destinations.